Update default dependencies for rpm packages.
In openSUSE Tumbleweed, there is no libuuid package, only libuuid1. In Fedora, there is no libuuid1 package, only libuuid. As RPM packages do not support a "package A OR package B" dependency, specifying a package name will inevitably result in a package that's either not installable on openSUSE or not installable on Fedora. As both libuuid1 on openSUSE and libuuid on Fedora install `/usr/lib64/libuuid.so.1` (at least on x86-64 machines), specifying a dependency on the file will allow the resulting RPM package to be installed on both openSUSE and Fedora.